feat: 스낵바 중복 표시 방지 로직 추가
- `SnackBarManager`에서 동일한 스낵바 요청이 단시간 내에 중복으로 발생하는 것을 방지하는 로직을 추가
